% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINDS-package.R
\docType{package}
\name{MINDS-package}
\alias{MINDS}
\alias{MINDS-package}
\title{MINDS: Bayesian Integrative Clustering For Mixed-Type Multimodal Data}
\description{
Fits the MINDS model, a Bayesian hierarchical mixture that clusters subjects
jointly on a binary and a continuous modality via a shared low-dimensional
latent construct space, using Polya-Gamma augmentation for the binary part.
Model selection over the number of clusters \code{Nc} and constructs \code{Nt} uses the
complete-data DIC family of Celeux et al. (2006).
}
\author{
\strong{Maintainer}: Yinjun Zhao \email{yz3503@cumc.columbia.edu}

}
\keyword{internal}
