% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minds_utils.R
\name{permn}
\alias{permn}
\title{List All Permutations}
\usage{
permn(n)
}
\arguments{
\item{n}{Integer; number of elements.}
}
\value{
A list of permutations, each as an integer vector.
}
\description{
List All Permutations
}
