% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minds_utils.R
\name{Zcat_to_Zest}
\alias{Zcat_to_Zest}
\title{Convert Cluster Labels to Binary Matrix}
\usage{
Zcat_to_Zest(Z.cat.est, Nc, Nb)
}
\arguments{
\item{Z.cat.est}{Integer vector of cluster assignments.}

\item{Nc}{Integer; number of clusters.}

\item{Nb}{Integer; number of subjects.}
}
\value{
An Nc x Nb binary matrix where \code{Z.est[k, i] = 1} if subject i is in cluster k.
}
\description{
Converts categorical cluster assignments to a binary membership matrix.
}
