\name{data_mixed}
\alias{data_mixed}
\docType{data}
\title{Example Simulated Mixed-Modality Dataset}
\format{A list with two elements:\cr
\describe{
  \item{y_1}{Binary matrix with subjects in rows and binary items in columns.}
  \item{y_2}{Continuous matrix with subjects in rows and continuous items in columns.}
}}
\usage{data(data_mixed)}
\description{Simulated mixed binary-continuous dataset for MINDS examples.}
